fix(utils.model.content): fix isValid setup for AbstractList and -Matrix
This is a workaround for the isValid method since generics could not be
used in this context.
This does however create the warning
"Content is a raw type. References to generic type Content<T> should be
parameterized"
and requires class casting in the subclasses so that the isValid method
receives the correct generic type of input.